app/requirements.txt
azure-devops
logzero
requests
GitPython
Office365-REST-Python-Client
debugpy
